Add C interface
This splits dettrace up so that it can be run via a C function call.
This makes it easier to pass callback functions to dettrace for
execution upon each system call. Loading a .so library is rather clunky,
so that has been removed.

The spawning of threads for /dev/[u]random has also been
opportunistically refactored such that it no longer requires global
static variables.