Support `injectSystemCall`
Add `injectSystemCall` so that the tracer could inject arbitrary syscalls for a
given tracee. The injected syscall will run synchronously, hence the tracer
should take care not to cause starving for the tracee being injected.

The return value is not processed, i.e. the tracer should do below check for
the return value:

    long retval = injectSystemCall(..., SYS_xxx, ...);
    if (retval < 0) {
        int this_errno = (int)-retval;  // errno, note we cannot update `__errno_location` for the tracee.
    } else {
        // snip
    }